style(documents): move file, update styling
- remove documents h1
- update loading state display
- move display-files to /components/Documents
- update svg links with proper icons
