+++ src/app/documents/page.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 
 import Navigation from '@/app/components/Navigation/Navigation';
-import DisplayFiles from '@/app/components/DisplayFiles';
+import DisplayFiles from '@/app/components/Documents/DisplayFiles';
 import RecentFiles from '@/app/components/Documents/RecentFiles';
 import UploadFile from '@/app/components/UploadFile';
 import FileView from '@/app/components/FileView';
@@ -216,7 +216,9 @@ export default function Documents() {
             {triggerUpload && (
                 <UploadFile triggerUpload={triggerUpload} setTriggerUpload={setTriggerUpload} />
             )}
-            {selectedLink && <LinkView link={selectedLink} onClose={handleCloseLinkView} />}
+            {selectedLink && (
+                <LinkView link={selectedLink} onClose={handleCloseLinkView} />
+            )}
             <div className={`flex-1 transition-all duration-200
                     ${isExpanded ? "pl-64" : "pl-12"}
                     ${selectedFile || triggerUpload || selectedLink ? "pr-96" : ""}
@@ -326,7 +328,6 @@ export default function Documents() {
                     </ul>
                     */}
                     <RecentFiles />
-                    <h1 className='text-2xl font-semibold my-4'>Documents</h1>
                     <DisplayFiles onFileClick={handleFileClick} onLinkClick={handleLinkClick} />
                 </div>
             </div>
